move posix_spawn test to wait
