+++ tests/process/wait.rs
@@ -1,5 +1,5 @@
 use libc::{kill, SIGSTOP};
-use rustix::process;
+use rustix::{io, process};
 use serial_test::serial;
 use std::process::{Command, Stdio};
 
@@ -40,4 +40,28 @@ fn test_wait() {
         .unwrap();
     assert!(status.stopped());
     assert_eq!(child_pid, pid);
+}
+
+#[test]
+#[serial]
+fn test_posix_spawn() {
+    let (read_pipe, write_pipe) = io::pipe().unwrap();
+    let message = "posix_spawn works";
+    let env_vars: &[&str] = &[];
+    let mut config = process::SpawnConfig::default();
+    let stdout = unsafe { io::stdout() };
+
+    config.add_dup2_action(&write_pipe, &stdout);
+
+    let pid =
+        process::posix_spawn("/usr/bin/echo", &["echo", "-n", message], env_vars, &config).unwrap();
+
+    // ensure reading from the pipe ends when the child finishes writing
+    core::mem::drop(write_pipe);
+    let mut buf = [0; 32];
+
+    let len = io::read(&read_pipe, &mut buf).unwrap();
+    let output = std::str::from_utf8(&buf[..len]).unwrap();
+    assert_eq!(output, message);
+    process::waitpid(Some(pid), process::WaitOptions::empty()).unwrap();
 }
